# cmd/api/docs/swagger.yaml
definitions:
...
  req.Coupons:
    properties:
      code:
        type: string
      discountmaximumamount:
        type: number
      discountpercent:
        type: number
      expirationdate:
        type: string
      minimumpurchaseamount:
        type: number
    type: object
...
  req.UpdateCoupon:
    properties:
      discountmaximumamount:
        type: number
      discountpercent:
        type: number
      expirationdate:
        type: string
      minimumpurchaseamount:
        type: number
    type: object
...
  res.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      stastuscode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/admin/coupon:
    get:
      consumes:
      - application/json
      description: Admins and users can see all available coupons
      operationId: view-coupons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admins and users can see all available coupons
      tags:
      - Coupon
  /api/admin/coupon/add:
    post:
      consumes:
      - application/json
      description: Admin can create new coupons
      operationId: create-coupon
      parameters:
      - description: details of new coupon to be created
        in: body
        name: new_coupon_details
        required: true
        schema:
          $ref: '#/definitions/req.Coupons'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Admin can create new coupon
      tags:
      - Coupon
  /api/admin/coupon/update/{couponId}:
    patch:
      consumes:
      - application/json
      description: admin coupon update
      operationId: UpdateCoupon
      parameters:
      - description: Coupon ID
        in: path
        name: couponId
        required: true
        type: string
      - description: New Admin details
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/req.UpdateCoupon'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update a existing coupon
      tags:
      - Coupon
...
  /api/user/coupon/apply/{code}:
    post:
      consumes:
      - application/json
      description: user can use his coupon code in cart
      operationId: apply-coupon
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: user apply coupon
      tags:
      - Coupon
  /api/user/coupon/remove:
    delete:
      consumes:
      - application/json
      description: user can remove the apply code in cart
      operationId: remove-coupon
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
      summary: user remove coupon
      tags:
      - Coupon
...
swagger: "2.0"